# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the SimpleTaskManager (STM) Server. See the GitHub repo '/doc/api/' for further details on authentication, websockets and changelogs.",
        "title": "SimpleTaskManager Server",
        "contact": {
            "name": "STM issue tracker",
            "url": "https://github.com/hauke96/simple-task-manager/issues"
        },
        "license": {
            "name": "GNU General Public License 3.0",
            "url": "https://github.com/hauke96/simple-task-manager/blob/master/LICENSE"
        },
        "version": "1.3.0"
    },
    "paths": {
...
        "/oauth2/callback": {
            "get": {
                "description": "OAuth2 callback called after OSM login. Performs the OAuth authentication by getting an OSM access token.",
                "tags": [
                    "authentication"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "The state-string that was given to the OAuth service.",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The authentication code to retrieve the access token.",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}